Bugfix in Taxonomy and Taxon copy and move functions.
This is still the same bug we tried to fix since a couple of commits. It is a nasty thing to work with blank pointers in a C++11-world.

This time, we hopefully got them all! The problem is that whenever the contents of the children_ container change, all contained elements need an update. There are many places where this is necessary.