Reimplement consensus sequence with threshold.
After conversation with the author of Seaview, it became clear that the approach "Use the ambiguity code for all occuring sites if the threshold is not reached" does not make too much sense. Thus, instead, we now add up frequencies until the threshold is reached, and use the ambiguity code of all used characters.