Change CMakeLists.txt to generate static and shared targets always
This change conditionally sets the property `EXCLUDE_FROM_ALL` on the
shared and static versions of the library, based on the options
`GENESIS_BUILD_SHARED_LIB` and `GENESIS_BUILD_STATIC_LIB`. The purpose
of this is so that apps which want to link with a specific version don't
have to bother with configuring anything, and instead can just link with
the shared or static target.