Fix gcc 7 optimization bug.
For some reason, gcc 7 tries to optimize a function kmer canonical encoding too much,
resulting in a warning that gets turned into an error:

error: assuming signed overflow does not occur when assuming that (X + c) >= X is always true [-Werror=strict-overflow]
         if( l + 2 >= k )

for the case in encode_gapped_palindrome_() where l==k.
So here we instead set l to k now, in the hope that this fixes the
issue.